Remove hard-coded sender address from the SMTP chat
Even though my name is indeed very famous, there is obviously no need to
broadcast it to the whole world :). The problem here is that the user enters the
address as a loosely-formatted string which is not trivial to parse. Let's stick
with a more generic stub as a crude hotfix.