Don't return unexpected data when lazy loading
The bug was that we returned a valid QVariant of a string/bytearray even when
asked for SizeHintRole. Such a QVariant is not invalid, but converts to zero
when casting to int, which confused the attached view.

We weren't hitting this bug because we set setUniformRowHeight in all views for
increased performance.

Reported by "th" on IRC.