Do not break with interleaved LIST responses
Previously we always flushed all untagged LIST replies in the queue upon
processing the tagged response for the LIST command. This caused issues with
IMAP servers which return interleaved sets if we feed them with pipelined
commands.

The only weird thing is that I did not catch this earlier, as I do see that on
Dovecot 1.1.7 on my laptop.