Clarify EmbeddedWebView's purpose
Previously, we were trying to communicate outside the minimal required size of
the viewport in order to make scrolling and scrollbars sane enough. This does
not seem to be required anymore with Qt 4.6.

OTOH, the class was doing plenty of other stuff (like disabling QWebView's
context menu) which should've been mentioned in the Doxygen.