Implement safe loading of external entities
When the currently displayed e-mail requires external network access, the user
is prompted for permission in an non-intrusive way. If she agrees, we enable
loading of such elements for current message and ask the message view to reload
stuff.

Still unimplemented are features like showing a list of URLs of things to load
etc. Also see comments in the added files.