Add some stub for handling the mailto: URL scheme internally
Note that Qt is currently (4.6.1) performing a case-sensitive comparison when
using a custom URL scheme handler; that needs fixing.

Also note that the QDesktopServices::setUrlHandler() does *not* support the
SLOT() macro at all, as of 4.6.1.  Debugging that was not really
straightforward, especially with a crying child just a few seat rows away.