Don't use O(n^2) when updating message UIDs
The problem was that TreeitemMessage::row() called QList::indexOf(), which has,
by definition, linear complexity. This was in turn called for each message in
mailbox, and that made things a bit slow on a mailbox with 40k messages (not as
slow as with some mainstream MUAs, but slow nonetheless, and it showed up in the
GUI latency).

Having tens of thousands of mailboxes is not so common and messages typically do
not have so many message parts at one level, so the only critical place where
performanced suffered badly was TreeItemMessage (mailboxes with huge amount of
messages are common). The implementation is a bit ugly, we have to re-implement
something which we got for free before.

This commit won't bring any performance improvement yet, we still check the data
against TreeItem::row() at this point.