Actually enable the performance boost with TreeItemMessage::row()
Callgrind suggests that we managed to cut the time usage of
QList<TreeItem*>::indexOf() from roughly 57% of the whole runtime to less than
1.3%, which is pretty cool (measured against a Dovecot instance with a
gentoo-dev ML archive; the slow test was over a local LAN, the other one over a
rather slow wifi, which might or might now have affected the result). The
important point here is that the GUI is much more responsive now.