Fix a really nasty segfault
In short, we were using QList::iterator and calling functions which could
(indirectly, through signals/slots, for the increased fun factor) modify the
original list using QList::append(). Because append() could invalidate iterators
(which is rather counter-intuitive, but a result of a possible memory
reallocation, and thus clearly possible), our previous code was not safe.

It's great to have valgrind at one's disposal. Without it, it would take much
more time to track it down.