Warn users before sending e-mails
The main reason behind this is that now that Trojita is getting tested by more
poeple, some of them might want to send real, "production" e-mails from it.
Given that that part is rather incomplete (for example, it doesn't deal with
Message-Id and friends at all), it's better to give a warning than to pollute
the Internet.