Fix some memleaks and make sure we eliminate some false positives
Even though the QEvent::DeferredDelete is marked as deprecated, it's what the
QCoreApplication's event loop delivers, and therefore something which we should
deliver, too. If we don't and we don't start the main event loop by ugly means
(see the removed bit of code) either, we'd get a memleak, which isn't
particulary nice behavior of a unit test, is it.

Oh, and there are more memleaks to be fixed :)