Add a special-case command for UID SEARCH command
The reason for this special case is that this command often accepts
weird-looking argument like a sequence set, and those do not really work well
with our heuristic to finding out proper encapsulation format. Formerly, they'd
be packed as quoted strings, which is clearly stupid; the correct way is to send
them as atoms.