Make sure to mark the current task as the active one in its perform()
This should actually be the very first thing the Task performs. If it calls
completed() before registering in the list of active tasks, it won't get
deleted, and the KeepMailboxOpen would not die, preventing its replacement by a
task responsible for different mailbox.

Also added some asserts around the Parser* to prevent littering the
Model::_parsers with possible zeros etc.