Make sure that we track IDLE completion correctly
This issue was reported by Bron Gondwana of the Cyrus fame.

Previously, we'd track the "is the IDLE still on the wire?" by tracking command
completion via its tagged response. However, if we combine that with pipelining
and checking the "are we in IDLE" upon each command queueing, there's a window
for races if the server is slow enough to respond to the DONE, which would
result in a DONE for each queued client Task, clearly violating the protocol
specification.

This should really be covered by unit tests...