wip: be careful with EXISTS response routing
This was reported by Bron Gondwana, tracking down the issue resulted in adding a
nice Task routing support.

The code still needs a betetr check, though -- the EXIST could arrive between we
queue a SELECT and the server receives it, effectively changing the number of
messages. That's what the [CLOSED] response code from RFC 5162 is for, in cases
the server sends that. And it would be nice not to break if the server doesn't
send it, perhaps revert back to not relying on condstore/qresync, even when
trojita will support it.