Queue failed messages after a timeout
This is required because of local cache artefacts -- when the cache contains
usable data for a mailbox state, Model will create messages with valid UIDs, and
these will be requested by the MailSynchronizer. However, when there's any
activity going on in the background which is blocking the Parser (and that's
pretty likely to happen, because we use just one connection at this time), the
requests for these messages won't be satisfied in a timely manner, leading to
timeouts.

Obviously, a good long-term solution is to move the rate limiting to
KeepMailboxOpenTask...