Initial blurb on delayed threading
When the source model changes and then settles down, at first we have to remove
existing rows, switch to the flat threading and then queue a request for real
threading.

This is the first step towards that code; right now, we use that "fake"
threading which builds a tree nested as much as possible.