It turned out that tracking items just by UID is not enough
The (parsed) THREAD response contains a zero as a placeholder for each message
which is known to "be there" but somehow missing from the mailbox. Our
oversimplified approach where we consider 0 to refer to a parent of all
top-level messages obvioulsy fails here, ocnsider the following thread:

    [a]
     + [0]
        + [b]
        + [c]
    [d]

This would get parsed as a flat list of (a, b, c, d).

So it looks like we will have to introduce a layer of indirection between the
internalId() of our QModelIndexes and the UIDs, just to support these "fake" or
"empty" messages.

(I can only imagine the fun in implementing yet another proxy model which would
simply hide these fake messages...)