XtConnect: use PgSQL-like command line options for the xtconnect-trojita
This allows the xtconnect-trojita app to use the PostgreSQL environment
variables or command line options to collect connection information. I've tried
to use the PG conventions for the command line options:
 -h hostname
 -p portnumber
 -d databasename
 -U username
 -W prompt for a password

The password prompt is sent to standard output and the password is read from
standard input. If -W has been passed and -U has not, the app also prompts for
and accepts the username.