Do not request individual flag updates
The underlying code was completely ready for batched updates, so there's really
no point in not using it -- it was even harder to do that individually :).

Thanks to Bron for reporting that a *long* time ago.

There's still no heuristic for delaying the updates a tiny bit, but I don't
think what would be worth the effort.

fixes #41