Batched loading of message metadata
Implement a queue of messages to fetch metadata for. This is very similar to the
queue already present for the message parts, with a difference that we do not
delay the request to the IMAP server, but queue it immediately after re-entering
the event loop. The goal here is to limit needless delays, while allowing for
higher efficiency in the case of "expensive access".

In short, this change will alsmot always fetch metadata for all messages which
are visible on the current screen using just one IMAP command.