Tasks: use the usual parent-child relations for KeepMailboxOpenTask and ObtainSynchronizedMailboxTask
Previously, the KeepMailboxOpenTask and ObtainSynchronizedMailboxTask were using
local ad-hoc way of communicating their progress for driving the
synchronization.  This had the usual drawback of "being special" and hard to
debug, but also meant that the tasks visualization would not work properly for
them.

This patch changes the basic behavior of KeepMailboxOpenTask; now they *always*
depend on the ObtainSynchronizedMailboxTask in the usual manner.  The
KeepMailboxOpenTask also keeps track of the waiting
ObtainSynchronizedMailboxTask, not the KeepMailboxOpenTask instances.