Do not special-case part fetching for message/rfc822
It appears that this got introduced in 65694b1c2f40ae477d9610237b71f7029b7ab462
(which predates the concept of the irregular shape of the TreeItemPart nesting).
The visible effect of the previous code is that a tooltip shown for the full
tree model would list just the header data.  That's cool, but it also has a big
disadwantage that the TreeItemPart's fetching status would remain at _loading,
and that the data would never hit the cache (acutally it gets stored, but with
the correct .HEADER suffix, and therefore it would get requested again next
time).

The existing GUI code of course uses the correct OFFSET_HEADER indexing which
means that this is not an issue at all.