Do not fetch parts which are already available from the cache
The "regular" code in the Gui called fetchFromCache() because it had to know
whether the item was already available.  This had the side effect of not
requesting the data over and over again, masking this bug.  However, in certain
situations (like the header view of a message/rfc822), this led to requesting
part data over network instead of simply going to the cache.