tests: Temporarily disable threading tests which work with a subset of messages
This is unfortunately a side effect of the recent additions to the threading
model that aim at reducing the number of "useless" THREAD commands. When the
received THREAD response doesn't contain the highest UID from the mailbox, we
will request the same data again and again.

This will have to be fixed somehow for processing of the search results and also
to mitigate possible issues when the server responds with smaller amount of
data.

refs #10