Tree: use lazy loading for message's special parts
Valgrind's massif pointed out that we're wasting quite a lot of space in there.
Even a dumb look at ImapModelThreadingTest::testThreadingPerformance (with a
modification to use 500k messages instead of mere 100k) suggests saving at least
180MB of RSS memory.

Note that we do not do the same for the special members of the TreeItemPart at
this point; they aren't allocated until the BODYSTRUCTURE arrives anyway.