Flags: change back from QSet<QString> to QStringList to save memory
Measurements [1] indicate that the heap memory usage grew from 296 MB to 385 MB
as a result of the switch to QSet.  The actual deduplication, as performed in
Model::normalizeFlags, appears to make sense, though (memory usage stays at 186 MB).

[1] ImapModelObtainSynchronizedMailboxTest::testFlagReSyncBenchmark with the
number of messages changed to 500k from default 100k. Measured on a 64bit Linux
with Qt 4.7.1.