IMAP: UID syncing shall happen immediately after receiving the UID SEARCH response
This has a huge benefit of fixing possible race conditions triggered by
arrivals/expunges processed while the sync is in progress -- basically, it's the
old time of check - time of use problem where the UID SEARCH is valid at the
time we receive the data, but not necessarily when the tagged OK for that
arrives.

Some tests are still disabled, though.