Add support for raw IMAP searching
This is based on an idea (and patch!) by Thomas Lübking. Unfortunately, due to
the (broken, arguably) API of the Parser's searching functions (which is
replicated at various other places throughout the whole stack), one has to add a
crude hack to Parser::searchHelper which makes sure that the data is sent
properly encoded.

It won't work completely reliably at this point, one cannot search for quotation
marks without LITERAL+ etc; that'd require a proper data structure to be passed
around instead of a QStringList.

That's something to be solved later.

refs #59
refs #528