Make the SubtreeModel "generic"
This is an extremely ugly hack. The problems it is trying to solve are:

* Proxy models cannot call createIndex on the parent model unless they are
  friends
* Qt's MOC doesn't support class templates
* Qt's MOC doesn't expand macros, either

So this is so far the "best way" I can come up with sharing a single
implementation of the code for creating subtree proxies. Coments welcome.

Oh, and all that is required because QML's frakkin VisualDataModel doesn't
support dynamic updates at all and passes stale QModelIndex instances back to
the model. That's even more sad.