Composer: fix assert crashes in drag-and-drop
It's an extremely wrong behavior to assert upon receving a drag-and-drop from
another application; a much clearer way is to validate the passed data and
politely refuse to accept the garbage.

Thomas Lübking has also suggested that it is wrong to fix the QDataStream
version at just one end of the pipe; he's right.