threading: be sure to report all changed columns
When a message deep in a thread changes, it might affect its parent. Care has to
be taken to make sure that we announce a change in all columns of the parent,
not just column zero.

The QTreeView actually doesn't break even when not signalled, but let's play it
nice and respect the expected contract.