threading: Prevent showing items with wrong font
This was actually pretty involved -- the main issue was that the flags were not
always loaded and that the subsequent updates were not being made available via
dataChanged.

Now, the lazy flag loading is actually good, so this change simply makes sure
that if the flags are actually needed, they will be requested.

fixes #494