xtuple: request related data for messages at once
Previsouly, the code would request data for message's headers and raw body at
once, even before the message metadata became available. However, because the
data is stored into the DB only after everything is available and the actual
main body part is determined only after the message metadata arrive, this has an
unfortunate effect of keeping essentially the whole mailbox in memory until each
and every message has been fetched. Needless to say, this led to a performance
which left much to be desired.

This patch changes the behavior so that we always fetch the tripple of (body,
headers, main part) at once. Unless the IMAP server plays ugly tricks with us
(like postponing the responses indefinitely), this leads to significant memory
savings because the memory can be freed almost immediately.