xtuple: don't waste time with these signals under xtconnect
Profiling shows that cca. 70% of time is spent in Model::parent due to calls to
TreeItem::row(). These calls are already optimized for the TreeItemMessage
(where they actually matter), so they are caused by the other classes. Because
the XtConnect is a heavy user of persistent indexes and because these require
significant amounts of calls to QAIM::parent, let's see if simply disabling
these notifications improves performance.