IMAP: don't assert crash on receiving BAD upon the initial conneciton setup
I've received an "y0 BAD Error in IMAP command received by server." when sitting
on some random heavily proxied network (the ZLUTY bus for those in .cz). Hitting
an assert as a result of a network response in clearly wrong, such a condition
shall be handled gracefully, even if by an early disconnect.