l10n: fix the output of older lconvert
Burkhard Lück has reported that my extraction sucks. It looks like the
differences between a .pot file which gets created on my machine (Qt 4.8.3) and
the scripty-generated one in SVN is that:

- "msgctxt" is replaced by the "#. ts-context" construct (and put into a weird place)
- "#, fuzzy" gets added
- "#, qt-format" strings are removed
- file names are different and are printed strictly on a one-per-line basis
- QML files are not processed

The first two issues are taken care of by this change.

I tried to do this in sed, but my head started spinning so fast when I read
about the loops that I gave up and went back to good old Python. Another day,
maybe.