l10n: produce a proper .po file header so that contexts are preserved again
This will have to be picked up by the translators, otherwise the contexts will
be all wrong and nothing will translated actually.

I hate this stuff. I'm *so* tempted with just moving the message extraction off
the KDE's scripty and its old lconvert, replacing it with a wget downloading my
own version. But maybe, jyst maybe this last thing is enough to make this beast
work, finally.

Meh.