IMAP: try to work with broken IMAP servers which do not report the UIDNEXT
Thanks to Jacopo De Simoi (wilder on Freenode) for a bugreport and a protocol
trace.

Please note that this is a rather bad situation, the missing UIDNEXT is treated
in the same way as a changed UIDVALIDITY. This *might* be relaxed in future, but
I quite frankly don't trust servers which violate this rule to not change the
immutable data for old UIDs, so this behavior is the best I can come up with for
now.

This means that Trojita will suck hard when talking to Courier.