Abandon LightMessageDataBundle
The whole point of this class was to prevent the tiny overhead of one uint.
Given that this class is only used in an uncommon case where user doesn't want
the perisstent caching and in the unit tests, it isn't worth the extra code
(which contained an obvious bug which was only uncovered by the coverity scan,
but was obscure enough to not be hit that often in the wild).