GUI: fix segfault in the FindBar dismissal
Because the underlying QWebView can be deleted more or less any time (for
example as a result of the user navigating away to another message), it is
really bad idea to keep it around indefinitely.

There was a reproducible segfault:
- open a message
- press ctrl+f, type something to execute a search
- navigate to another message
- dismiss the find bar

Nothing more besides using a guarded pointer is actually needed here; the checks
appear to be placed in all entry points already.