build: Move the git version to an extra library for linking speedups
Previously, the git version handling was a part of Common. However, this
resulted in relinking each time the Common.a changed, which was on every build;
it led to a needless relink of all unit tests all the time.

This change forces the relink only for the Gui and Harmattan projects which
actually use the gitversion feature.

Also move the code to a proper namespace.