IMAP: make GMail happy when invoking ESEARCH
There's currently a bug in GMail's implementation of the ESEARCH command where
they don't correctly handle the SEARCH command with an empty return
specification ("RETURN ()") which the RFC mandates to be equivalent to "RETURN
(ALL)".

I've already reported that to Google and Brandon has promised a fix. Our users
can also add ESEARCH to the blacklisted capabilities in order to make it work,
but this is far from being easily discoverable. Let's waste these three bytes
until GMail rolls out a release with a proper fix -- this is a major
interoperability issue which prevents even basic mailbox resynchronization, so
it's worth the effort IMHO.