IMAP: accept NIL when looking for unsigned ints
Thanks to Ashley Penney <apenney@gmail.com> (ashp on IRC) for reporting.

This is apparently required for broken IMAP servers like the "davmail" gateway
to Exchange(tm). Getting NIL is roughly as bad as -1 or -2 which we already
handle, so there's no reason for complaying so loudly in this case.