qmake: replace custom system with a boring qmake setup
Antti Poikela <antti.poikela@iki.fi> (amppa on IRC) reported that trojita failed
to link with mimetypes-qt4 on Windows. This is due to the path manipulation
which Trojita's own linking.pri performs on Windows. That path manipulation is
in turn required because of qmake not tracking dependencies on static libraries
by default (at least that's how it worked back in 2009). Let's just throw away
the complex and non-standard setup of mimetypes-qt4 with something which might
have a chance of working on Windows.

Also complain loudly when the QStandardPaths emulation is not available --
that's the best we can do instead of failing later during build.