Composer: implement a well-behaving progress bar
Sadly, the IMAP code doesn't keep us informed about the progress of the APPEND
operation. Better than showing a progress bar stuck at, say, 10%, we show a
fuzzy one meaning "yeah, we're working on it". When the APPEND finishes, the
progress moves to 50% and from then on, tracks the actual progress as reported
by the MSA doing the actual submission.