Rework the drag & drop in the message list view to only skip UID=0 messages
The whole stack assumed that a message can only be drag-and-droped after it has
been loaded. That's clearly nonsense, the only real requirement is that the
knowing the UID is enough.

This was rather annoying when selecting messages via Ctrl+A on a testing
instance with plenty of messages and no persistent cache (and hence 45k+
messages not loaded which could not be moved via Ctrl+A).

The new code has a nice side effect of always rendering the "loading-UID-not-known"
messages as disabled, greyed-out entities, which is also kind of cool.

Thanks to Stefan de Konink <stefan@konink.de> (skinkie on IRC) and Brian Jackson
<iggy@theiggy.com> (iggy on IRC) for their reports about slow drag-and-drop
(which is not fixed yet).