GUI: validate each address separately and also when filled in programatically
Previously, the validation was performed based on the contents of all items.
This meant that after a single field contained invalid data, any subsequent
modification to any other field would render the newly modified field in red.
That's wrong, the validation shall be based exclusively on data from the
modified field.

This patch also triggers the validator when setting the data programatically.

CCMAIL: sreepriya1111@gmail.com