Addressbook: improve handling of prettyName when creating new contact
If prettyName is empty, use the default value for the name field (so
don't show an empty name field, users will think they cannot click
there).

If prettyName is not empty, set the text of the QStandardItem to
prettyName. Before, the text of the QStandardItem would be
"[New Contact]" until we selected another contact.