cmake: Add the cmake-based setup
This is squashed from many commits, hence the huge blob.

The answers at
http://stackoverflow.com/questions/3780667/use-cmake-to-get-build-time-svn-revision
were more than inspiring when fixing the gitversion thing.

Thanks to Caspar who reported a failure when no .po files are around, the
Debian's linker's nature when it comes to linking with non-listed DSOs, etc.

The minimal required CMake version is 2.8.7 for Qt4 builds (that's what gets
shipped on the oldest supported Ubuntu). When building Qt5, we have to use 2.8.9
or newer, otherwise the qt5_use_modules macro is not available.

The build setup for the Harmattan version was shamelessly stolen from KDE's work
-- huge thanks to those who committed that to their repositories.