SMTP: do not signal success after hitting a failure earlier on
OK, this is some ugly code -- but it appears that it makes QwwSmtpClient more or
less work. That class is in a bad need for replacement :(.

The problem was that the onDisconnected event handler within the QwwSmtpClient
simply checked whether there are any pending commands, and if there are none, it
proceeded further towards emitting success even though the connection has never
succeeded. This is a wrong assumption in the client's code -- signalling success
just because the connection was severed before any command had a chance to be
queued is not what I would expect when using such component.

This code does *not* prevent duplicate signals about success, but I have yet to
see a situation where it breaks something.

BUG: 321272
REVIEW: 111073