Unify handling of "open attachment" and "download attachment"
The point of this change is to make sure that we don't leak objects when
anything wrong happens during attachment download/opening/whatever. To that end,
the following changes were made:

- The lifetime of both FileDownloadManager instances is now unified; they are
allocated only when the corresponding action is about to take place, saving
memory
- The GUI buttons (QActions, actually) are disabled immediately when the user
clicks them -- this is to prevent multiple requests of the same action
- The FileDownloadManager no longer signals the time it starts transferring
data. Instead, there's a new signal telling the rest of the code that the action
was cancelled, i.e. no data were written.