GUI: Composer: sync the action state with the current selection harder
Looks like the slots connected to currentChanged and rowsInserted are not
enough; it was clearly visible when dropping message/rfc822 attachments to the
view -- the checkbox for "Send inline" was not selected.

Switching to custom context menu processing effectively fixes this problem as we
now have precise control on action state.

I'm still leaving these slots in place because it makes sense to manipulate the
action state anyway -- suppose there was e.g. a "Delete" key assigned to
removing attachments...