Composer: move the "Re: " prefix to the start of the subject
While I personally like the "[foo] [bar] Re: meh" a lot, it looks like no other
MUA is doing this. Let's stick with the crowd and switch to putting the Re: at
the beginning of the line.

The sanitization is still in place, though -- duplicate Re: stanzas or mailing
list prefixes are removed.

REVIEW: 112002