HTML: the regexps for detecting links work on top of *escaped* data
A regression was introduced in 6e56db69. Previously, the HTML escaping was done
via custom placeholders and the regexps therefore had to work on more or less
raw, unescaped data. Since the switch to Qt's own HTML escaping, the regular
expressions see the already-escaped form, and therefore the logic for matching
the special characters have to be changed so that it looks for "&amp;" instead
of "&" ("<", ">" and '"' are not used in these regexps and Qt does not escape
apostrophes ("'")).

Thanks to Caspar for report and for his patch. This commit takes a different
approach without introducing custom "magic strings".

BUG: 323390
REVIEW: 112024