tests: MIME: Check support of "language" within the "encoding" of the encoded-words
RFC2231 extended [1] the rules for RFC2047-style encoded words so that they can
include an optional language specification withing the used "encoding"
declaration. Our MIME parser has supported this from the very beginning, but it
was not explicitly tested.

Thanks to Gnome's Jeffrey Stedfast for blogging about the fun of MIME decoding
[2] which pointed to gmime's unit tests [3] which made me check this
conformance.

On a quick scan, it appears that we support most of their MIME compatibility
hacks (i.e. hacks which support them to read broken input). We do not support
encoded-words which are split in the middle of a multibyte character (and I've
seen such mails in the fly).

[1] http://tools.ietf.org/html/rfc2231, section 5
[2]
http://jeffreystedfast.blogspot.cz/2013/08/why-decoding-rfc2047-encoded-headers-is.html
[3] https://git.gnome.org/browse/gmime/tree/gmime/gmime-utils.c#n1723