GUI: MessageView: show an indicator whenever the message parts are being loaded
It looks that QWebView does not have support for showing any placeholder
whenever it is loading data from network; instead, a simple empty widget is
shown. This is suboptimal because the user doesn't know that the mail is
currently loading unless they pay close attention to the global
application-level progress indicator in the status bar.

This patch propagates loadingStarted/loadingFinished signals of the enclosed
SimplePartWidget instances to the MessageView which simply shows/hides another
progress indicator.

Another idea was to implement this on the QNAM level, but it turned out that the
same MsgPartNetAccessManager is used also for attachment download which is not
what we are interested in.

REVIEW: 112092