Add libPlugins library for implementing and using addressbook and password plugins
jkt: This commit also removes the -fPIE flag which is no longer necessary --
we're now forcing -fPIC on all objects. Without this change, the
trojita+plugins.so wouldn't link and fail due to relocations.

jkt(2): Be sure to update the build recipes for OBS. This change adds a new
target and it is important to respect the platform's customs about the proper
placement.

REVIEW: 112275