GUI: fix segfault due to uninitialized variable
LineEdit has two constructors and one of them was missing a proper
initialization of the m_historyEnabled. This broke an assumption that a
(m_historyEnabled being true) implies (completer() != 0). One could trigger this
by opening the composer, attempting to send a mail with an empty subject with
cursor in the recipients widget, and pressing Esc twice which led to calling
QCompleter::completionMode with "this" set to nullptr.