GUI: control text/plain vs. text/html preference via PartLoadingOptions
The work on the Kontact plugin requires a different way of accessing the user's
preferences because the QSettings's default constructor no longer works as
expected. Towards that end, Pali is working on having the QSettings access
abstracted away, and this patch makes it possible to avoid passing the
QSettings* around.

REVIEW: 112155